+++ rollup.config.js
@@ -1,0 +1,16 @@
+import commonjs from "@rollup/plugin-commonjs";
+import { nodeResolve } from "@rollup/plugin-node-resolve";
+import json from '@rollup/plugin-json';
+
+const config = {
+  input: "./index.js",
+  output: {
+    esModule: true,
+    file: "dist/index.js",
+    format: "es",
+    sourcemap: true,
+  },
+  plugins: [commonjs(), nodeResolve({ preferBuiltins: true }), json()],
+};
+
+export default config;