+++ CHANGELOG.md
@@ -1,9 +1,13 @@
+## 0.0.3 (2025-08-04)
+
+- allow using a custom `labels.json` file.
+
 ## 0.0.2 (2025-03-14)
 ### Bug fixes
 
-- use `labels.json` from action repository
+- use `labels.json` from action repository.
 
 ## 0.0.1 (2025-03-12)
 ### Features
 
-- first release of prototype action
+- first release of prototype action.